Revert "Implement Multi Language Support (#2906)"
This reverts commit abf97571caaa69722e0b1244857b6493a21cb741.
