+++ app/[slug]/client-page.tsx
