+++ content/whats-new-tinacloud/2025.02.92.mdx
@@ -1,0 +1,12 @@
+---
+versionNumber: 2025.02.92
+dateReleased: 2025-02-26T05:44:04Z
+---
+
+## What's Changed
+### 🔧 Fixes
+* 🔊 Logs - Improves structured logging for the Authorizer by @ncn-ssw
+* 🥅 Errors - Return HTTP 404 for GraphQL queries on missing branches by @ncn-ssw
+### 👒 Dependencies
+* chore(deps): bump koa from 2.15.3 to 2.15.4 by @dependabot
+
