+++ content/whats-new-tinacms/v2.7.1.mdx
@@ -1,0 +1,11 @@
+---
+versionNumber: 2.7.1
+dateReleased: 2025-02-26T06:01:23Z
+---
+
+### Patch Changes
+
+-   [#5517](https://github.com/tinacms/tinacms/pull/5517) [`74513b3`](https://github.com/tinacms/tinacms/commit/74513b357aa27165aa86f7b3218c697c663539e8) Thanks [@kldavis4](https://github.com/kldavis4)! - Fix bug when encoding media manager image urls with special characters
+
+-   [#5509](https://github.com/tinacms/tinacms/pull/5509) [`06c1716`](https://github.com/tinacms/tinacms/commit/06c17163b558a96275b0ef66c746b005a6f90d13) Thanks [@Ben0189](https://github.com/Ben0189)! - add auto format for strikethrough
+
