+++ tina/collectionsSchema/pages.tsx
@@ -89,14 +89,11 @@ export const pagesCollection = {
   path: 'content/blocksPages',
   format: 'json',
   ui: {
-    router: ({ document, collection }) => {
-      const fullPath = document._sys.path;
-      const basePath = `${collection.path}`;
-      const relativePath = fullPath
-        .substring(basePath.length)
-        .replace(/\.[^/.]+$/, '');
-
-      return `${relativePath}`;
+    router: ({ document }) => {
+      if (document._sys.filename === 'home') {
+        return `/`;
+      }
+      return `/${document._sys.filename}`;
     },
     beforeSubmit: async ({
       values,
